1task: теперь установка позиции работает правильно

+++ 1-lab/task1/Letter/Position.h
@@ -2,6 +2,6 @@
 
 struct Position
 {
-    int m_x;
-    int m_y;
+    float m_x;
+    float m_y;
 };
